requirements.txt
streamlit==1.28.1
pillow==10.0.1
requests==2.31.0
google-cloud-vision==3.4.4
openai==1.3.0
anthropic==0.7.0
groq==0.4.0
python-dotenv==1.0.0
pandas==2.1.3
plotly==5.17.0
numpy==1.24.3
google-generativeai==0.3.2 
imgurpython==1.1.7 